Added view-cheese-by-category feature
CONTROLLER - Added handler method-"category" mapped to "cheese/category", it che
cks through iterable of cheeses via CheeseDao object, filters them by the given
category and adds the matches to an arraylist of cheeses which is passed into th
e view.

VIEWS - The controller directs to "cheese/index.html" with a title reflecting th
e category.

Links to this view are added to /categories and to /cheese/index